requirements.txt
flickrapi
python-dotenv
